Random Picker: split the tool into audio / spin-hook / wheel modules
RandomPickerTool.tsx was carrying four jobs. Now, following the calls/
colocated-parts pattern: audio.ts (SpinSound: WebAudio tick/ding/mute),
useSpinWheel.ts (spin lifecycle: winner pick, rAF rotation-following,
transitionend settle), Wheel.tsx (presentational SVG), and the tool file
is just composition + the form. No behaviour change.
